requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.7
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
Pillow==10.0.0